refactor(api): decouple FastAPI config from route handlers
- Move agent-instances routes to routes/agent_instances.py
- Move tasks routes to routes/tasks.py
- Keep FastAPI app config, lifespan, and /health in main.py
- Include routers in main.py using app.include_router()

Closes #28
