+++ src/server/api/routes/__init__.py
@@ -1,0 +1,6 @@
+"""API route handlers."""
+
+from src.server.api.routes.agent_instances import router as agent_instances_router
+from src.server.api.routes.tasks import router as tasks_router
+
+__all__ = ["agent_instances_router", "tasks_router"]
